+++ src/bach/test/series.mjs
@@ -1,3 +1,5 @@
+import { disableDeprecationWarnings, resetDeprecationFlags } from '../../deprecation.mjs';
+
 const { default: expect } = await import('expect');
 
 const { default: bach } = await import('../index.js');
@@ -23,6 +25,10 @@ function fnError(done) {
 }
 
 describe('bach: series', function () {
+  beforeEach(() => {
+    disableDeprecationWarnings();
+    resetDeprecationFlags();
+  });
   it('should can take the sync functions without return value', function (done) {
     bach.series([fnSync])(function (error, results) {
       expect(error).toEqual(null);