+++ src/run-function.js
@@ -1,15 +1,21 @@
 const { customPromisify } = require('./custom-promisify.js');
 
-const cacheEnabled = false;
 const cache = new Map();
 async function call(fn) {
-  if (cacheEnabled && cache.has(fn)) {
-    return cache.get(fn);
+  const deprecations = await import('./deprecation.mjs');
+  const runTaskOnceFlag = deprecations.getTaskRunsOnceDeprecationFlag();
+
+  if (cache.has(fn)) {
+    if (!runTaskOnceFlag) {
+      deprecations.warn(
+        `[DEPRECATED] The task ${fn.name} has been called more than once. This is deprecated. Please visit https://github.com/XaveScor/gulp/blob/master/DEPRECATIONS.md for more information`,
+      );
+    } else {
+      return cache.get(fn);
+    }
   }
   const result = await customPromisify(fn)();
-  if (cacheEnabled) {
-    cache.set(fn, result);
-  }
+  cache.set(fn, result);
 
   return result;
 }